// backend/src/errors/error-codes.enum.ts
export enum ErrorCode {
  // Bad Request [400]
  MISSING_IMAGE = 'MISSING_IMAGE',
  MISSING_FILE = 'MISSING_FILE',
  MISSING_EMAIL = 'MISSING_EMAIL',
  MISSING_PASSWORD = 'MISSING_PASSWORD',
  MISSING_VERIFICATION_CODE = 'MISSING_VERIFICATION_CODE',

  //Unauthorized [401]
  ONLY_ADMINS_ALLOWED = 'ONLY_ADMINS_ALLOWED',

  //Forbidden [403]
  INVALID_ID = 'INVALID_ID',
  UNABLE_TO_DELETE_THIS_IMAGE = 'UNABLE_TO_DELETE_THIS_IMAGE',
  UNABLE_TO_DELETE_THIS_COMMENT = 'UNABLE_TO_DELETE_THIS_COMMENT',
  UNABLE_TO_EDIT_THIS_COMMENT = 'UNABLE_TO_EDIT_THIS_COMMENT',
  UNABLE_TO_DELETE_THIS_REPLY = 'UNABLE_TO_DELETE_THIS_REPLY',
  UNABLE_TO_EDIT_THIS_POST = 'UNABLE_TO_EDIT_THIS_POST',
  UNABLE_TO_DELETE_THIS_POST = 'UNABLE_TO_DELETE_THIS_POST',
  INCORRECT_CREDENTIALS = 'INCORRECT_CREDENTIALS ',
  USER_SENT_INVALID_VERIFICATION_CODE = 'USER_SENT_INVALID_VERIFICATION_CODE',

  //Not Found Errors [404]
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  POST_NOT_FOUND = 'POST_NOT_FOUND',
  IMAGE_NOT_FOUND = 'IMAGE_NOT_FOUND',
  COMMENT_NOT_FOUND = 'COMMENT_NOT_FOUND',
  VERIFICATION_CODE_NOT_FOUND = 'VERIFICATION_CODE_NOT_FOUND',
  REPLY_NOT_FOUND = 'REPLY_NOT_FOUND',
  USER_HAVE_NOT_LIKED_THIS_POST = 'USER_HAVE_NOT_LIKED_THIS_POST',
  USER_HAVE_NOT_DISLIKED_THIS_POST = 'USER_HAVE_NOT_DISLIKED_THIS_POST',
  USER_HAVE_NOT_LIKED_THIS_COMMENT = 'USER_HAVE_NOT_LIKED_THIS_COMMENT',
  USER_HAVE_NOT_DISLIKED_THIS_COMMENT = 'USER_HAVE_NOT_DISLIKED_THIS_COMMENT',
  USER_HAVE_NOT_LIKED_THIS_REPLY = 'USER_HAVE_NOT_LIKED_THIS_REPLY',
  USER_HAVE_NOT_DISLIKED_THIS_REPLY = 'USER_HAVE_NOT_DISLIKED_THIS_REPLY',

  //Conflict [409]
  IMAGE_ALREADY_EXIST = 'IMAGE_ALREADY_EXIST',
  USER_ALREADY_LIKED_THIS_POST = 'USER_ALREADY_LIKED_THIS_POST',
  USER_ALREADY_LIKED_THIS_COMMENT = 'USER_ALREADY_LIKED_THIS_COMMENT',
  USER_ALREADY_DISLIKED_THIS_COMMENT = 'USER_ALREADY_DISLIKED_THIS_COMMENT',
  USER_ALREADY_LIKED_THIS_REPLY = 'USER_ALREADY_LIKED_THIS_REPLY',
  USER_ALREADY_DISLIKED_THIS_REPLY = 'USER_ALREADY_DISLIKED_THIS_REPLY',
  EMAIL_IS_ALREADY_IN_USE = 'EMAIL_IS_ALREADY_IN_USE',
  USERNAME_IS_ALREADY_IN_USE = 'USERNAME_IS_ALREADY_IN_USE',

  //Unsupported Media [415]
  FILE_NOT_SUPPORTED = 'FILE_NOT_SUPPORTED',
}